requirements.txt

streamlit
deepagents
langchain
langchain-openai
langchain-google-genai
langchain-groq
langchain-nvidia-ai-endpoints
langchain-community
langchain-chroma
openai
pymupdf
python-docx
chromadb
protobuf==3.20.3
tavily-python
exa-py
googlemaps
python-dotenv
python-multipart
pydantic
pandas
tinydb
fastapi
uvicorn


# uv pip install -r requirements.txt
